HowToの記述、PlayTimeの追加

+++ Game/Game/src/Battle.cpp
@@ -155,7 +155,7 @@ void Battle::update()
 
 void Battle::draw() const
 {
-	Rect(0, Scene::Height() * 0.7, Scene::Width(), Scene::Height() * 0.3)
+	Rect(0, (int32)(Scene::Height() * 0.7), Scene::Width(), (int32)(Scene::Height() * 0.3))
 		.draw(Arg::top = ColorF(0.0, 0.0), Arg::bottom = ColorF(0.0, 0.5));
 
 	// スコアの表示
